Changed all occurences of the "Graphics" Object to "Graphics2D". TODO: -The LifeBars need to be fixed,       -The images on ImageComponents need to be cached as scaling them on the fly results in a significant performance loss.
git-svn-id: https://dev.wibe.media/svn/code/java/trunk/litiengine@1571 16493bb7-3577-b145-b9c3-72a97d279fb6

+++ src/de/gurkenlabs/litiengine/graphics/IGuiComponent.java
@@ -9,28 +9,30 @@ public interface IGuiComponent extends IRenderable {
    *
    * @return the height
    */
-  public int getHeight();
+  public double getHeight();
 
   /**
    * Gets the width.
    *
    * @return the width
    */
-  public int getWidth();
+  public double getWidth();
 
   /**
    * Gets the x.
    *
    * @return the x
    */
-  public int getX();
+  public double getX();
 
   /**
    * Gets the y.
    *
    * @return the y
    */
-  public int getY();
+  public double getY();
+  
+ 
 
   /**
    * Prepare.
@@ -46,8 +48,8 @@ public interface IGuiComponent extends IRenderable {
    */
   public void suspend();
 
-  void setPosition(int x, int y);
+  void setPosition(double x, double y);
   void setPosition(Point2D newPosition);
 
-  void setDimension(int width, int height);
+  void setDimension(double width, double height);
 }
